requirements.txt
fastapi==0.79.0
bcrypt==3.2.2
hypercorn==0.13.2
passlib==1.7.4
tortoise-orm==0.19.2
pyjwt==2.4.0
python-decouple==3.6